requirements.txt
google-generativeai==0.8.4
python-dotenv>=1.0.1
pytesseract>=0.3.10
mss>=9.0.1
Pillow>=10.0.0